+++ lib/src/services/releases_service/current_release_parser.dart
@@ -1,10 +1,5 @@
 import 'dart:io';
 
-const _arm = 'arm';
-const _arm64 = 'arm64';
-const _ia32 = 'ia32';
-const _x64 = 'x64';
-
 /// Parsed release model
 class ParsedReleases {
   /// Constructor
@@ -52,28 +47,4 @@ ParsedReleases parseCurrentReleases(Map<String, dynamic> json) {
     channels: currentRelease,
     releases: releases,
   );
-}
-
-// https://stackoverflow.com/questions/45125516/possible-values-for-uname-m
-final _unames = {
-  'arm': _arm,
-  'arm64': _arm64,
-  'aarch64_be': _arm64,
-  'aarch64': _arm64,
-  'armv8b': _arm64,
-  'armv8l': _arm64,
-  'i386': _ia32,
-  'i686': _ia32,
-  'x86_64': _x64,
-};
-
-String _architecture() {
-  final uname = Process.runSync('uname', ['-m']).stdout as String;
-  final trimmedName = uname.trim();
-  final architecture = _unames[trimmedName];
-  if (architecture == null) {
-    throw Exception('Unrecognized architecture: "$trimmedName"');
-  }
-
-  return architecture;
 }