+++ lib/src/utils/fvm_client.dart
@@ -62,21 +62,6 @@ class FVMClient {
   /// Returns a list of projects by providing a list of [directories]
   static final fetchProjects = ProjectService.fetchProjects;
 
-  /// Returns true if [cacheVersion] is configured as global
-  static final checkIfGlobal = CacheService.isGlobal;
-
-  /// Returns true if FVM global version is configured corretly
-  static final checkIfGlobalConfigured = CacheService.isGlobalConfigured;
-
-  /// Returns sync a global configured cache version
-  static final getGlobalVersionSync = CacheService.getGlobalVersionSync;
-
-  /// Returns async a global configured cache version
-  static final getGlobal = CacheService.getGlobal;
-
-  /// Sets a [version] as global
-  static final setGlobalVersion = CacheService.setGlobal;
-
   /// Scans for Flutter projects found in the rootDir
   static final scanDirectory = ProjectService.scanDirectory;