+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 3.0.0
 
 - Faster SDK downloads! Local Git Flutter repository caching is now enabled by default.
-- Global version has been deprecated. For more info go [here](https://fvm.app/docs/guides/global_version)
+- Global version has been deprecated. For more info go [here](https://fvm.app/docs/guides/global_version).
 - Removed skip-setup configuration.
 
 ## 2.4.1